+++ spec/serializers/service_serializer_spec.rb
@@ -1,7 +1,21 @@
 require 'spec_helper'
 
 describe ServiceSerializer do
+
   let(:service_model) { Service.new }
+
+  let(:image_status) do
+    double(:image_status,
+      info: {
+        'Config' => {
+          'ExposedPorts' => {'3000/tcp' => {} }
+        }
+      })
+  end
+
+  before do
+    Docker::Image.stub(:get).and_return(image_status)
+  end
 
   it 'exposes the attributes to be jsonified' do
     serialized = described_class.new(service_model).as_json
@@ -23,7 +37,8 @@ describe ServiceSerializer do
       :sub_state,
       :type,
       :errors,
-      :docker_status
+      :docker_status,
+      :default_exposed_ports
     ]
     expect(serialized.keys).to match_array expected_keys
   end